Minor improvements
The most important change is that the plugin does not break if the
`plugins.afterBrunch` key is missing in the config. For example, it lets
you use afterBrunch only in the production override.

The check in the constructor for whether the commands exist or not is
removed. Missing commands we're run anyway, causing "command does not
exist" messages to show up twice.

stdout and stderr are now printed. Especially stderr is useful, since
tools can tell you about errors in your commands.

A small error in the example code in the readme is corrected. The code is
also syntax highlighted now.

The sentence about the `plugins` object of the config is removed, since
that is how plugins are supposed to provide config.

A useful npm prepublish hook is added.

The `onCompile` method no longer returns an unnecessary array.